% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAnalysisDetails.R
\name{createAnalysesDetails}
\alias{createAnalysesDetails}
\title{createAnalysesDetails}
\usage{
createAnalysesDetails(
  outputFolder,
  removeSubjectsWithPriorOutcome = TRUE,
  asUnitTest = FALSE
)
}
\value{
File with format 'cmAnalysisList' for CohortMethod
}
\description{
A function for generating the varying cross products of PS models to fit for analysis
}
\examples{
createAnalysesDetails(outputFolder)

}
