% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultDataModel.R
\name{uploadResultsToDatabase}
\alias{uploadResultsToDatabase}
\title{Upload results to database}
\usage{
uploadResultsToDatabase(
  connectionDetails,
  schema,
  resultsFolder,
  tablePrefix = "",
  forceOverWriteOfSpecifications = FALSE,
  purgeSiteDataBeforeUploading = FALSE,
  databaseIdentifierFile = "cdm_source_info.csv",
  runCheckAndFixCommands = TRUE,
  warnOnMissingTable = TRUE,
  purgeDataModel = TRUE
)
}
\description{
Upload results to database
}
